requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
dateparser==1.1.8
dnspython==2.3.0
frozenlist==1.3.3
hmmlearn==0.2.8
idna==3.4
joblib==1.2.0
motor==3.1.1
multidict==6.0.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1675642512762/work
pandas==1.5.3
pycryptodome==3.17
pymongo==4.3.3
python-binance==1.0.17
python-dateutil==2.8.2
python-telegram-bot==13.15
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
TA-Lib @ file:///home/conda/feedstock_root/build_artifacts/ta-lib_1655667894234/work
threadpoolctl==3.1.0
tornado==6.1
tzdata==2023.2
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.15
websockets==10.4
yarl==1.8.2
